business-idea-github-bundle/requirements.txt
gradio>=4.37.0
python-dotenv>=1.0
pandas>=2.0
numpy>=1.23
requests>=2.31
textblob>=0.18
yfinance>=0.2
